+++ FullStack_Projects/EduHub_Project/services/IUserService.cs
@@ -4,10 +4,10 @@ namespace EduHub_Project.services
 {
     public interface IUserService
     {
-        Task<User> GetUserAsync(long userId);
+        Task<User> GetUserAsync(int userId);
         Task<IEnumerable<User>> GetAllStudentsAsync();
         Task<User> CreateUserAsync(User newUser);
-        Task<User> UpdateUserAsync(long userId, User updatedUser);
-        Task<bool> DeleteUserAsync(long userId);
+        Task<User> UpdateUserAsync(int userId, User updatedUser);
+        Task<bool> DeleteUserAsync(int userId);
     }
 }
